clinical_trial_env/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "clinical-trial-env"
version = "1.0.0"
description = "ClinicalTrialEnv — Adaptive clinical trial design RL environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "pydantic>=2.6.0",
    "requests>=2.31.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27.0"]
